+++ __main__.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Entry point for python -m cam
+Entry point for python -m cammy
 """
 
 import sys
 
-# Import from src/cam/cli.py
-from src.cam.cli import main
+# Import from src/cammy/cli.py
+from src.cammy.cli import main
 
 if __name__ == "__main__":
     sys.exit(main())