+++ src/cammy/__main__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Entry point for python -m cam
+Entry point for python -m cammy
 """
 
 import sys